criteria package is moved.
criteria package is moved to under util package.
