+++ api/src/main/java/ca/bc/gov/educ/api/course/repository/CourseCriteriaQueryRepository.java
@@ -3,7 +3,7 @@ package ca.bc.gov.educ.api.course.repository;
 import org.springframework.stereotype.Repository;
 
 import ca.bc.gov.educ.api.course.model.entity.CourseEntity;
-import ca.bc.gov.educ.api.course.repository.criteria.CriteriaQueryRepository;
+import ca.bc.gov.educ.api.course.util.criteria.CriteriaQueryRepository;
 
 @Repository
 public interface CourseCriteriaQueryRepository extends CriteriaQueryRepository<CourseEntity> {
