+++ api/src/main/java/ca/bc/gov/educ/api/course/repository/CourseCriteriaQueryRepositoryImpl.java
@@ -3,7 +3,7 @@ package ca.bc.gov.educ.api.course.repository;
 import org.springframework.stereotype.Repository;
 
 import ca.bc.gov.educ.api.course.model.entity.CourseEntity;
-import ca.bc.gov.educ.api.course.repository.criteria.CriteriaQueryRepositoryImpl;
+import ca.bc.gov.educ.api.course.util.criteria.CriteriaQueryRepositoryImpl;
 
 @Repository
 public class CourseCriteriaQueryRepositoryImpl extends CriteriaQueryRepositoryImpl<CourseEntity> implements CourseCriteriaQueryRepository {
